Split state/mod.rs into submodules
